To fix issue #5812 I moved the responsibility for apply a stylesheet to child elements to the Element class.
It would be weird for a stylesheet to get all parent stylesheets of an element and call Apply on them too. I think it's a good thing that StyleSheet.Apply(Element) really only applies it on the given Element? The main issue was that when the ApplyParentStyleSheets was called, the stylesheet of the element itself wasn't applied, so you could never override parent style elements